requirements.txt
# Core ML and AI libraries
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
faiss-cpu>=1.7.4
numpy>=1.21.0
accelerate>=0.20.0

# MLX (Mac only - optional)
mlx>=0.4.0
mlx-vlm>=0.0.8

# Document processing
beautifulsoup4>=4.12.0
markdown>=3.4.0
pillow>=9.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Local LLM interface
ollama>=0.1.7

# Optional: Performance monitoring
psutil>=5.9.0